requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Deprecated==1.2.18
deprecation==2.1.0
fpdf==1.7.2
greenlet==3.1.1
idna==3.10
img2pdf==0.5.1
iniconfig==2.0.0
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
ocrmypdf==16.9.0
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20240706
pi_heif==0.21.0
pikepdf==9.5.2
pillow==11.1.0
playwright==1.49.1
pluggy==1.5.0
pycparser==2.22
pyee==12.0.0
Pygments==2.19.1
pytesseract==0.3.13
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.6.2
python-slugify==8.0.4
requests==2.32.3
rich==13.9.4
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
